completed auth and created rich text editor

+++ src/app/api/post/comment/delete/[delete]/route.ts
@@ -1,3 +1,4 @@
+import { auth } from "@/auth";
 import { getJwtTokenData } from "@/utils/getJwtTokenData";
 import { PrismaClient } from "@prisma/client";
 import { NextRequest, NextResponse } from "next/server";
@@ -9,17 +10,22 @@ export async function DELETE(
   { params }: { params: { delete: string } }
 ) {
   try {
-    getJwtTokenData(req);
-    const id = params.delete;
+    const session = await auth();
+    const userId = session?.user.id;
+    if (userId) {
+      const id = params.delete;
 
-    const deleteComment = await prisma.comments.delete({
-      where: { id: Number(id) },
-    });
+      const deleteComment = await prisma.comments.delete({
+        where: { id: Number(id) },
+      });
 
-    return NextResponse.json(
-      { message: "Deleted Successfully", success: true },
-      { status: 200 }
-    );
+      return NextResponse.json(
+        { message: "Deleted Successfully", success: true },
+        { status: 200 }
+      );
+    } else {
+      return null;
+    }
   } catch (error: any) {
     return NextResponse.json({ error: error.message }, { status: 500 });
   }
